Fix a bug that an URI is not resolved right.
There was an UriNotHierarchicalException during generating AspectJ code
because the URI has not been resolved right. This was due the fact that
since Tycho Dresden OCL dependencies are packed in jars, thus the file
within the jar needs to be extracted first to a temporary location.
Changin from FileLocator.resolve(URL) to FileLocator.toFileURL(URL)
fixed the problem.